[BE] feat: page, today page 경로 변경
page 경로 변경.
시퀄라이즈 핫픽스를 위해 커밋
