--- web/next.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  //   async redirects() {
-  //     return [
-  //       {
-  //         source: "/",
-  //         // destination: "/",
-  //         //   permanent: true,
-  //       },
-  //     ];
-  //   },
-};
