+++ web/pages/api/test.ts
@@ -1,0 +1,5 @@
+import { NextApiRequest, NextApiResponse } from 'next'
+
+export default function handler(req:NextApiRequest,res:NextApiResponse){
+    res.json({stage:process.env.STAGE})
+}
