+++ web/pages/api/users/index.ts
@@ -2,11 +2,11 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { sampleUserData } from '../../../utils/sample-data'
 
 const handler = (_req: NextApiRequest, res: NextApiResponse) => {
+  console.log(process.env.STAGE)
   try {
     if (!Array.isArray(sampleUserData)) {
       throw new Error('Cannot find user data')
     }
-
     res.status(200).json(sampleUserData)
   } catch (err) {
     res.status(500).json({ statusCode: 500, message: err.message })
